requirements.txt
pyautogen
pyautogen[gemini]
dotenv
